feat(website): Port the website to Next.js
Currently was using middleman app to do the frontend
managment. But Next.js is pretty easy and appealing.
So ported the website (basically github pages) over to Next.js.

With this blog part is also done as we can now write the markdown
versions of posts and compile it. (Was not able to do this is
in the middleman version.)